project-1/LLM_Playground/requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
